+++ rocon_gateway/src/rocon_gateway/gateway_node.py
@@ -17,7 +17,7 @@ from urlparse import urlparse
 # Local imports
 import hub_discovery
 import gateway
-import hub_api
+import hub_manager
 
 ##############################################################################
 # Gateway Configuration and Main Loop Class
@@ -36,7 +36,7 @@ class GatewayNode():
         self._param = rocon_gateway.setup_ros_parameters()
         key = uuid.uuid4()  # random 16 byte string, alternatively uuid.getnode() returns a hash based on the mac address, uuid.uid1() based on localhost and time
         self._unique_name = self._param['name'] + key.hex  # append a unique hex string
-        self._hub_manager = hub_api.HubManager(
+        self._hub_manager = hub_manager.HubManager(
                              hub_whitelist=self._param['hub_whitelist'],
                              hub_blacklist=self._param['hub_blacklist']
                              )
